requirements.txt
# Core runtime
pydantic>=2.6
python-dateutil>=2.8
pandas>=2.2
numpy>=1.26
pillow>=10.2
opencv-python>=4.9
faker>=24.0
reportlab>=4.1

# OCR (optional)
# paddleocr
# paddlepaddle

# VLM (optional)
# transformers
# accelerate
# qwen-vl-utils

# Reporting
jinja2>=3.1
rich>=13.7

# Testing
pytest>=8.0
